+++ hibernate4/src/main/java/com/fasterxml/jackson/datatype/hibernate4/HibernateProxySerializer.java
@@ -1,10 +1,7 @@
 package com.fasterxml.jackson.datatype.hibernate4;
 
 import java.io.IOException;
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.util.HashMap;
-import java.util.concurrent.ConcurrentHashMap;
 
 import org.hibernate.engine.spi.Mapping;
 import org.hibernate.engine.spi.SessionImplementor;
@@ -38,9 +35,9 @@ public class HibernateProxySerializer
 
     protected final boolean _forceLazyLoading;
     protected final boolean _serializeIdentifier;
-    private boolean _usePersistentClassForIdentifier;
+    protected final boolean _usePersistentClassForIdentifier;
     protected final Mapping _mapping;
-    static final ConcurrentHashMap<CacheKey, Field> _idFieldCache = new ConcurrentHashMap<HibernateProxySerializer.CacheKey, Field>();
+    protected final IdentifierTypeFactory _identifierTyperFactory;
 
     /**
      * For efficient serializer lookup, let's use this; most
@@ -75,6 +72,11 @@ public class HibernateProxySerializer
         _mapping = mapping;
         _dynamicSerializers = PropertySerializerMap.emptyMap();
         _property = null;
+        if(_usePersistentClassForIdentifier){
+            _identifierTyperFactory = new IdentifierTypeFactory();
+        }else {
+            _identifierTyperFactory = null;
+        }
     }
 
     /*
@@ -172,7 +174,7 @@ public class HibernateProxySerializer
                 final Object idValue = init.getIdentifier();
                 if (_usePersistentClassForIdentifier) {
                     Class<?> persistentClass = init.getPersistentClass();
-                    Object identifierClass = createTargetIdentifierClassAndSetIdValue(persistentClass, idName, idValue);
+                    Object identifierClass = _identifierTyperFactory.createInstanceWithIdValue(persistentClass, idName, idValue);
                     return identifierClass;
                 } else {
                     HashMap<String, Object> map = new HashMap<String, Object>();
@@ -186,100 +188,9 @@ public class HibernateProxySerializer
     }
 
 
-    Object createTargetIdentifierClassAndSetIdValue(Class<?> persistentClass, final String idName, final Object idValue) {
-        try {
-            Constructor<?> defaultConstructor = persistentClass.getDeclaredConstructor();
-            defaultConstructor.setAccessible(true);
-            Object instance = defaultConstructor.newInstance();
-            Field idField = getIdField(persistentClass, idName, instance);
-            if(!idField.isAccessible()){
-                idField.setAccessible(true);
-            }
-            idField.set(instance, idValue);
-            return instance;
-        } catch (Exception e) {
-            throw new IllegalStateException("Error creating identifier class [" + persistentClass.getSimpleName() + "] and "
-                    + "setting idValue with [" + idName + "=" + idValue + "]. Default constructor present and field available?",
-                    e);
-        }
-    }
-
-    private Field getIdField(Class<?> persistentClass, String idName, Object instance) throws NoSuchMethodException {
-        CacheKey cacheKey = new CacheKey(persistentClass.getCanonicalName(), idName);
-        Field field = _idFieldCache.get(cacheKey);
-        if (field != null) {
-            return field;
-        }
-        else {
-            synchronized (_idFieldCache) {
-                field = _idFieldCache.get(cacheKey);
-                if (field == null) {
-                    field = findField(persistentClass, idName, instance);
-                    _idFieldCache.put(cacheKey, field);
-                }
-                return field;
-            }
-        }
-    }
-
-    private Field findField(Class<?> persistentClass, String idName, Object instance) throws NoSuchMethodException {
-        if (persistentClass == null) {
-            return null;
-        }
-        Field field;
-        try {
-            field = persistentClass.getDeclaredField(idName);
-        } catch (NoSuchFieldException e) {
-            field = findField(persistentClass.getSuperclass(), idName, instance);
-        }
-        return field;
-    }
-
-    private static class CacheKey {
-        private final String clazz;
-        private final String idName;
-
-        public CacheKey(String clazz, String idName) {
-            super();
-            this.clazz = clazz;
-            this.idName = idName;
-        }
-
-        @Override
-        public int hashCode() {
-            final int prime = 31;
-            int result = 1;
-            result = prime * result + ((clazz == null) ? 0 : clazz.hashCode());
-            result = prime * result + ((idName == null) ? 0 : idName.hashCode());
-            return result;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (this == obj)
-                return true;
-            if (obj == null)
-                return false;
-            if (getClass() != obj.getClass())
-                return false;
-            CacheKey other = (CacheKey) obj;
-            if (clazz == null) {
-                if (other.clazz != null)
-                    return false;
-            } else if (!clazz.equals(other.clazz))
-                return false;
-            if (idName == null) {
-                if (other.idName != null)
-                    return false;
-            } else if (!idName.equals(other.idName))
-                return false;
-            return true;
-        }
-
-    }
-    
+        
     public static void clearIdFieldCache(){
-        _idFieldCache.clear();
+        IdentifierTypeFactory._idFieldCache.clear();
     }
 
 }
