+++ hibernate4/src/test/java/com/fasterxml/jackson/datatype/hibernate4/IdentifierTypeFactoryTest.java
@@ -5,16 +5,16 @@ import java.util.Map;
 import org.junit.Assert;
 import org.junit.Test;
 
-public class HibernateProxySerializerTest {
+public class IdentifierTypeFactoryTest {
 
     @Test
-    public void testCreateTargetIdentifierClassAndSetIdValue() throws Exception {
-        HibernateProxySerializer s = new HibernateProxySerializer(false);
-        Object object = s.createTargetIdentifierClassAndSetIdValue(TestClassA.class, "id", 1L);
+    public void testCreateInstanceWithIdValue() throws Exception {
+        IdentifierTypeFactory s = new IdentifierTypeFactory();
+        Object object = s.createInstanceWithIdValue(TestClassA.class, "id", 1L);
         Assert.assertEquals(TestClassA.class, object.getClass());
         Assert.assertEquals(((SuperClass) object).id.longValue(), 1L);
         @SuppressWarnings("rawtypes")
-        Map cache = HibernateProxySerializer._idFieldCache;
+        Map cache = IdentifierTypeFactory._idFieldCache;
         Assert.assertTrue(cache.size() == 1);
     }
